refac: Clean up after medallion package merge (#243)
