+++ source/core/src/core/bronze/infrastructure/streams/bronze_repository.py
@@ -1,6 +1,6 @@
 from pyspark.sql import DataFrame, SparkSession
 
-from core.bronze.domain.constants import BronzeTableNames
+from core.bronze.infrastructure.config import BronzeTableNames
 from core.settings.catalog_settings import CatalogSettings
