+++ source/core/src/core/bronze/infrastructure/streams/kafka_stream.py
@@ -1,7 +1,6 @@
 from pyspark.sql import DataFrame, SparkSession
 
-import core.bronze.infrastructure.helpers.path_helper as path_helper
-from core.bronze.domain.constants import BronzeTableNames
+from core.bronze.infrastructure.config import BronzeTableNames
 from core.bronze.infrastructure.config.storage_container_names import StorageContainerNames
 from core.bronze.infrastructure.settings import (
     KafkaAuthenticationSettings,
@@ -9,6 +8,7 @@ from core.bronze.infrastructure.settings import (
     SubmittedTransactionsStreamSettings,
 )
 from core.settings.catalog_settings import CatalogSettings
+from core.utility.shared_helpers import get_checkpoint_path
 
 
 class KafkaStream:
@@ -20,7 +20,7 @@ class KafkaStream:
         self.bronze_database_name = CatalogSettings().bronze_database_name  # type: ignore
 
     def submit_transactions(self, spark: SparkSession) -> None:
-        checkpoint_location = path_helper.get_checkpoint_path(
+        checkpoint_location = get_checkpoint_path(
             self.data_lake_settings, StorageContainerNames.bronze, "submitted_transactions"
         )
         write_stream = (
@@ -42,7 +42,7 @@ class KafkaStream:
         self,
         dataframe: DataFrame,
     ):
-        checkpoint_location = path_helper.get_checkpoint_path(
+        checkpoint_location = get_checkpoint_path(
             self.data_lake_settings, StorageContainerNames.bronze, "processed_transactions"
         )
         event_hub_instance = KafkaAuthenticationSettings().event_hub_instance  # type: ignore
