+++ source/core/src/core/gold/infrastructure/adapters/delta_gold_adapter.py
@@ -3,14 +3,10 @@ from typing import Callable
 from pyspark.sql import DataFrame
 
 from core.gold.application.ports.gold_port import GoldPort
-from core.gold.infrastructure.config.database_names import DatabaseNames
 from core.gold.infrastructure.config.storage_container_names import StorageContainerNames
-from core.gold.infrastructure.shared_helpers import (
-    EnvironmentVariable,
-    get_checkpoint_path,
-    get_env_variable_or_throw,
-    get_full_table_name,
-)
+from core.settings.catalog_settings import CatalogSettings
+from core.utility.environment_variable_helper import EnvironmentVariable, get_env_variable_or_throw
+from core.utility.shared_helpers import get_checkpoint_path, get_full_table_name
 
 
 class DeltaGoldAdapter(GoldPort):
@@ -37,4 +33,7 @@ class DeltaGoldAdapter(GoldPort):
             df_write_stream.start().awaitTermination()
 
     def append(self, df: DataFrame, table_name: str) -> None:
-        df.write.format("delta").mode("append").saveAsTable(get_full_table_name(DatabaseNames.gold, table_name))
+        catalog_settings = CatalogSettings()  # type: ignore
+        df.write.format("delta").mode("append").saveAsTable(
+            get_full_table_name(catalog_settings.gold_database_name, table_name)
+        )
