--- source/core/src/core/gold/infrastructure/config/database_names.py
@@ -1,3 +1,0 @@
-class DatabaseNames:
-    # Should probably be an environment variable because it is created in Terraform.
-    gold = "measurements_gold"
