--- source/core/src/core/silver/infrastructure/config/container_names.py
@@ -1,3 +1,0 @@
-class ContainerNames:
-    bronze = "bronze"
-    silver = "silver"
