+++ source/core/src/core/silver/infrastructure/config/storage_container_names.py
@@ -1,0 +1,2 @@
+class StorageContainerNames:
+    silver = "silver"
