+++ source/core/src/core/utility/environment_variable_helper.py
