+++ source/core/tests/utility/test_shared_helpers.py
@@ -1,9 +1,5 @@
-from unittest.mock import patch
-
-from core.gold.infrastructure.shared_helpers import (
-    EnvironmentVariable,
+from core.utility.shared_helpers import (
     get_checkpoint_path,
-    get_env_variable_or_throw,
     get_full_table_name,
     get_storage_base_path,
 )
@@ -47,32 +43,3 @@ def test__get_full_table_name__should_return_expected():
 
     # Assert
     assert result == expected
-
-
-@patch("os.getenv")
-def test__get_env_variable_or_throw_found__should_return_expected(mock_getenv):
-    # Arrange
-    mock_getenv.return_value = "testvalue"
-    variable = EnvironmentVariable.DATALAKE_STORAGE_ACCOUNT
-
-    # Act
-    result = get_env_variable_or_throw(variable)
-
-    # Assert
-    assert result == "testvalue"
-    mock_getenv.assert_called_once_with(variable.name)
-
-
-@patch("os.getenv")
-def test__get_env_variable_or_throw_not_found__should_throw_exception(mock_getenv):
-    # Arrange
-    mock_getenv.return_value = None
-    variable = EnvironmentVariable.DATALAKE_STORAGE_ACCOUNT
-
-    # Act & Assert
-    try:
-        get_env_variable_or_throw(variable)
-        assert False
-    except ValueError as e:
-        assert str(e) == f"Environment variable not found: {variable.name}"
-    mock_getenv.assert_called_once_with(variable.name)
